Agregar a Wish hecho, componente wish no funciona
